feat: put filters in their own file
